chore: rename Constraints file into Constraint
