+++ include/Sema/ConstraintSystem.hpp
@@ -1,0 +1,14 @@
+#ifndef GLU_SEMA_CONSTRAINT_SYSTEM_HPP
+#define GLU_SEMA_CONSTRAINT_SYSTEM_HPP
+
+#include "Constraints.hpp"
+
+namespace glu::sema {
+
+class ConstraintSystem {
+
+}
+
+}
+
+#endif // GLU_SEMA_CONSTRAINT_SYSTEM_HPP
