+++ lib/Sema/CMakeLists.txt
@@ -15,5 +15,5 @@ target_link_libraries(Sema
 target_sources(Sema
     PRIVATE
     GlobalScopeVisitor.cpp
-    Constraints.cpp
+    Constraint.cpp
 )
